requirements-dev.txt
annotated-types==0.6.0
anthropic==0.25.7
anyio==4.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
et-xmlfile==1.1.0
filelock==3.14.0
fsspec==2024.3.1
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
joblib==1.4.2
lxml==5.2.2
numpy==1.26.4
openai==1.33.0
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
pgvector==0.2.5
playwright==1.43.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
